renamed numba defaults class and moved it to top directory
