+++ src/pygama/math/functions/crystal_ball.py
@@ -9,8 +9,8 @@ import numpy as np
 from math import erf
 
 from pygama.math.functions.pygama_continuous import pygama_continuous 
-from ..utils import numba_math_defaults_kwargs as nb_kwargs
-from ..utils import numba_math_defaults as nb_defaults
+from pygama.utils import numba_math_defaults_kwargs as nb_kwargs
+from pygama.utils import numba_math_defaults as nb_defaults
