+++ src/pygama/math/functions/linear.py
@@ -6,8 +6,8 @@ import numpy as np
 from numba import prange
 
 from pygama.math.functions.pygama_continuous import pygama_continuous 
-from ..utils import numba_math_defaults_kwargs as nb_kwargs
-from ..utils import numba_math_defaults as nb_defaults
+from pygama.utils import numba_math_defaults_kwargs as nb_kwargs
+from pygama.utils import numba_math_defaults as nb_defaults
 
 
 @nb.njit(**nb_kwargs)
