+++ src/pygama/math/functions/moyal.py
@@ -10,8 +10,8 @@ from numba import prange
 from math import erfc
 
 from pygama.math.functions.pygama_continuous import pygama_continuous
-from ..utils import numba_math_defaults_kwargs as nb_kwargs
-from ..utils import numba_math_defaults as nb_defaults
+from pygama.utils import numba_math_defaults_kwargs as nb_kwargs
+from pygama.utils import numba_math_defaults as nb_defaults
